Updated animations, add test music add player sounds
