+++ GGJ21/Assets/Audio/SFX_Footstep_Hard_4.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 12616b650cbcb4bdcaa07e4b80b7db68
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
