+++ GGJ21/Assets/Audio/SFX_Footstep_Hard_6.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 78d9a7224e1b34ae1b6daaf13c111444
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
