--- GGJ21/Assets/Models.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 41e86a410a7329245942259fc91510ce
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
