+++ GGJ21/Assets/Scripts/AnimationStateController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5cfb62121260a4caf9e7c99ef710f2bd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
