+++ GGJ21/ProjectSettings/TagManager.asset
@@ -5,6 +5,7 @@ TagManager:
   serializedVersion: 2
   tags:
   - Hidden
+  - Floor
   layers:
   - Default
   - TransparentFX
